requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-dev.in
#
appdirs==1.4.4            # via black
attrs==20.2.0             # via pytest
backcall==0.2.0           # via ipython
black==20.8b1             # via -r requirements-dev.in
click==7.1.2              # via black
coverage==5.3             # via -r requirements-dev.in, pytest-cov
decorator==4.4.2          # via ipython, traitlets
factory-boy==3.1.0        # via -r requirements-dev.in
faker==4.4.0              # via factory-boy
flake8==3.8.4             # via -r requirements-dev.in
freezegun==1.0.0          # via -r requirements-dev.in
importlib-metadata==2.0.0  # via flake8, pluggy, pytest
iniconfig==1.0.1          # via pytest
ipython-genutils==0.2.0   # via traitlets
ipython==7.16.1           # via -r requirements-dev.in
jedi==0.17.2              # via ipython
mccabe==0.6.1             # via flake8
mypy-extensions==0.4.3    # via black
packaging==20.4           # via pytest
parso==0.7.1              # via jedi
pathspec==0.8.0           # via black
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.7     # via ipython
ptyprocess==0.6.0         # via pexpect
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via flake8
pyflakes==2.2.0           # via flake8
pygments==2.7.1           # via ipython
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.1        # via -r requirements-dev.in
pytest-django==3.10.0     # via -r requirements-dev.in
pytest==6.1.1             # via -r requirements-dev.in, pytest-cov, pytest-django
python-dateutil==2.8.1    # via faker, freezegun
regex==2020.9.27          # via black
six==1.15.0               # via packaging, python-dateutil, traitlets
text-unidecode==1.3       # via faker
toml==0.10.1              # via black, pytest
traitlets==4.3.3          # via ipython
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
wcwidth==0.2.5            # via prompt-toolkit
zipp==3.3.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
